Switch to declarative database schema
MAIN-1155
